requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
scikit-learn
rapidfuzz
passlib[argon2]
argon2-cffi
PyJWT
google-generativeai
pydantic[email]
numpy
pandas
